+++ openslides/utils/autoupdate.py
@@ -10,7 +10,8 @@ from mypy_extensions import TypedDict
 
 from .cache import element_cache, get_element_id
 from .projector import get_projector_data
-from .utils import get_model_from_collection_string, is_iterable, timeprint
+from .utils import get_model_from_collection_string, is_iterable
+from .timing import timeprint
 
 
 class AutoupdateElementBase(TypedDict):
@@ -75,7 +76,7 @@ class AutoupdateBundle:
         if not self.autoupdate_elements:
             return
 
-        a0 = time.time()
+        a = [time.time()]
 
         for collection, elements in self.autoupdate_elements.items():
             # Get all ids, that do not have a full_data key
@@ -94,18 +95,18 @@ class AutoupdateBundle:
                 for full_data in model_class.get_elements(ids):
                     elements[full_data["id"]]["full_data"] = full_data
 
-        a1 = time.time()
+        a.append(time.time())
 
         # Save histroy here using sync code.
         save_history(self.elements)
 
-        a2 = time.time()
+        a.append(time.time())
 
         # Update cache and send autoupdate using async code.
         async_to_sync(self.async_handle_collection_elements)()
 
-        a3 = time.time()
-        print("done(): 1:", a1-a0, "2:", a2-a1, "3:", a3-a2, "sum:", a3-a0)
+        a.append(time.time())
+        timeprint("done()", a)
 
     @property
     def elements(self) -> Iterable[AutoupdateElement]:
